requirements.txt
aiohttp==3.9.5
aiogram==3.6.0
python-dotenv==1.0.1
aiohttp-socks==0.8.4
motor==3.4.0